在代码中添加颁布勋章的代码

+++ medal-service/src/main/java/com/macaku/medal/domain/entry/HarvestAchievement.java
@@ -18,6 +18,6 @@ public class HarvestAchievement {
 
     private Long userId;
 
-    private Long degree;
+    private Integer degree;
 
 }
